requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
psycopg2-binary
python-multipart
